Update Class UML Diagram (Final)
Updated based on the instruction from the Professor provided below:                                            
 
Client class anything: Connection related, send, receive message. 
The ChatWindow class should only focus on displaying messages and managing the UI. 

You may want another class that handles user input on the ChatWindow or it could be handled by the ChatWindow directly.
 But the ChatWindow must not have any knowledge about the client connection.
Also, some methods are missing return types.


Last, compare the class diagram to the sequence diagram and make sure they both match. That is the methods shown on each are used in each other. 
The class diagram provides a static picture of the classes and their methods. The sequence diagram shows when the methods are used and the interactions between classes.